// src/generated/openapi/models/screenInput.ts
/**
 * Generated by orval v6.21.0 🍺
 * Do not edit manually.
 * screen-geometry-app-backend-serverless-apis-v1
 * OpenAPI spec version: 1.0.0
 */

/**
 * Screen primitive data
 */
export interface ScreenInput {
  aspectRatio: string
  darkColor: string
  diagonalSize: number
  hRes?: number
  lightColor: string
  vRes?: number
}